% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler}
\alias{profiler}
\title{Profiler Function}
\usage{
profiler(data, output_file = "profiler_report.md", sep = ",")
}
\arguments{
\item{data}{A data frame or a character string representing the file path to the data.}

\item{output_file}{A character string specifying the name of the output markdown file
where the profiler report will be saved. Default is "profiler_report.md".}

\item{sep}{A character string specifying the separator used if data is a file path, default is ",".}
}
\value{
A markdown report generated by the appropriate profiling function, either for data frames or files.
}
\description{
This function generates a profiler report for a given dataset or file.
It checks the type of input data and calls the appropriate profiling function
for data frames or files.
}
\examples{
# Example with a data frame
df <- data.frame(a = 1:5, b = letters[1:5])
profiler(df)

# Example with a file path
profiler("data/my_data.csv")

}
