% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler_dataframe}
\alias{profiler_dataframe}
\title{Generate Profiler Report from Data Frame}
\usage{
profiler_dataframe(data, output_file = "profiler_report.md")
}
\arguments{
\item{data}{A data frame for which the profiler report will be generated.}

\item{output_file}{A character string specifying the name of the output markdown file
where the profiler report will be saved. Default is "profiler_report.md".}
}
\value{
None. The function does not return a value but modifies the specified markdown file.
}
\description{
This function generates a profiler report for a given data frame by calculating
various statistics and writing them to a specified markdown file. It can also
generate visualizations for the variables, interactions, correlations, etc. in the data frame.
}
\details{
The function validates that the input is a data frame and that the output
file name is a character string. It generates an overview section in the report,
which currently includes dataset statistics, by calling the \code{write_statistics}
function. Further functionalities such as variable types and visualizations
can be added in the future.
}
\examples{
# Example usage with a data frame
df <- data.frame(a = 1:5, b = letters[1:5])
profiler_dataframe(df)

}
