% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler_file}
\alias{profiler_file}
\title{Read Data from File and Generate Profiler Report}
\usage{
profiler_file(file_path, output_file = "profiler_report.md", sep = ",")
}
\arguments{
\item{file_path}{A character string specifying the path to the file containing the data.
The file must exist. The function currently supports CSV files.}

\item{output_file}{A character string specifying the name of the output markdown file
where the profiler report will be saved. Default is "profiler_report.md".}

\item{sep}{A character string specifying the separator used in the file. Default is ",".}
}
\value{
None. The function does not return a value but modifies the specified markdown file
through the \code{profiler_dataframe} function.
}
\description{
This function reads a dataset from a specified file and generates a profiler report
by calling the \code{profiler_dataframe} function. It validates the input file path and
the output markdown file name before proceeding.
}
\details{
The function checks that the \code{file_path} is a character string and that the
specified file exists. It reads the data from the file and then passes the
resulting data frame to the \code{profiler_dataframe} function to generate the report.
}
\examples{
# Example usage with a CSV file
profiler_file("data/my_data.csv")

}
